requirements.txt
Jinja2
MarkupSafe
PyYAML